==> code/1-upload-raw-data.ipynb <==
import json
import numpy as np
from os import getcwd
import pandas as pd

def set_path():
   abspath = getcwd()
   dname = os.path.dirname(abspath)
   os.chdir(dname)
set_path()


# Open raw data
df_train = pd.read_json("data/raw/train.json")
df_train.to_pickle("data/processed/dirty_df.pkl")

df_test = pd.read_json("data/raw/test.json")
df_test.to_pickle("data/processed/test_dirty_df.pkl")
