==> code/3.6-model-3papers.ipynb <==
## In this file:
- Using data of 1702 frequent authors (having 3 or more than 3 papers in the data sets)
- Model: NB


import numpy as np
import pandas as pd
from os import getcwd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_selection import chi2
import os
import pickle
from collections import Counter
from sklearn import  naive_bayes, pipeline
from sklearn.metrics import accuracy_score


def set_path():
    abspath = getcwd()
    dname = os.path.dirname(abspath)
    os.chdir(dname)
set_path()

df_train = pd.read_pickle('data/processed/train_3pp_df.pkl')
df_val = pd.read_pickle('data/processed/val_3pp_df.pkl')
df_train.shape, df_val.shape


#### Bag of words:

Get x common words out of each author's corpus

# checking number of unique words in each author's corpus 
def get_author_unique_corpus(df):
   listauthor = df['authId_enc'].unique()
   corpus_auth = {}
   for au in listauthor:
      totalword = []
      auth_content = df.loc[df['authId_enc'] == au, 'content']
      for art in auth_content:
         wordlist = art.split()
         totalword += wordlist
      corpus_auth[au] = list(set(totalword))
   return corpus_auth
corpus_by_author = get_author_unique_corpus(df_train)

num_corpus = {k:len(v) for k, v in corpus_by_author.items() }
max_uniq_w = max(num_corpus, key = num_corpus.get)
print(f'author {max_uniq_w} has maximum unique words: {num_corpus[max_uniq_w]}' )
min_uniq_w = min(num_corpus, key = num_corpus.get)
print(f'author {min_uniq_w} has maximum unique words: {num_corpus[min_uniq_w]}' )
print(f'average number of unique words per author is {np.mean(list(num_corpus.values()))}' )




### Feature engineering and selection with Tfidf



vectorizer = TfidfVectorizer(max_features=3000, )
corpus = df_train['content']
vectorizer.fit(corpus)
Xtrain = vectorizer.transform(corpus)


# feature selection by chi2
Xnames = vectorizer.get_feature_names_out() ## original full corpus
p_value_limit = 0.95
df_features = pd.DataFrame()
for author in np.unique(df_train['authId_enc']):
   X= Xtrain.toarray()
   Y = df_train['authId_enc'] ==author
   chi, p = chi2(X, Y)
   df_features = pd.concat([df_features, 
   pd.DataFrame({'feature': Xnames, 'score' : 1-p, 'authId_enc': author})])

   df_features = df_features[df_features['score'] > p_value_limit]
df_features = df_features.sort_values(['authId_enc', 'score'],
                                          ascending = [True, False])
Xnames = df_features['feature'].unique().tolist() # selected words for new corpus from feature selection

count_uniq_w_author = Counter(df_features["authId_enc"])
min_auth = min(count_uniq_w_author, key = count_uniq_w_author.get)
print(f'author {min_auth} has the least unique chosen words {count_uniq_w_author[min_auth]} ')

# PRINT out selected feature by author
for author in np.unique(df_train['authId_enc'])[:10]:
   print("# {}:".format(author))
   print("  . selected features:",
         len(df_features[df_features["authId_enc"]==author]))
   print("  . top features:", ",".join(
df_features[df_features["authId_enc"]==author]["feature"].values[:10]))
   print(" ")     

len(Xnames)

### Fit vectorizer in the new corpus

vectorizer = TfidfVectorizer(vocabulary=Xnames)
vectorizer.fit(corpus)
Xtrain_new = vectorizer.transform(corpus)
ytrain = df_train['authId_enc'].values
yval= df_val['authId_enc'].values

classifierNB = naive_bayes.MultinomialNB()
## pipeline
modelNB = pipeline.Pipeline([("vectorizer", vectorizer),  
                           ("classifier", classifierNB)])
## train classifier
modelNB["classifier"].fit(Xtrain_new, ytrain)
## test
Xval = df_val['content'].values
predicted = modelNB.predict(Xval)
predicted_prob = modelNB.predict_proba(Xval)

accuracy = accuracy_score(yval, predicted)


accuracy

### Fine tune MultiNomial NB model with Gridsearch

Xtrain = df_train['content']
ytrain = df_train['authId_enc'].values
vectorizer = TfidfVectorizer()
classifierNB = naive_bayes.MultinomialNB()
modelNB = pipeline.Pipeline([("vectorizer", vectorizer),  
                           ("classifier", classifierNB)])

from sklearn.model_selection import GridSearchCV
grid_params = {
  'classifier__alpha': np.linspace(0.5, 1.5, 3),
  'classifier__fit_prior': [True, False],
  'vectorizer__max_df': [0.1, 0.5, 1],
  'vectorizer__binary': [True, False],
  'vectorizer__norm': [None, 'l1', 'l2'], 
  'vectorizer__max_features': [1000, 1500, 2000, 2500, 3000],
}
clf = GridSearchCV(modelNB, grid_params, cv = 2, scoring='accuracy')
clf.fit(Xtrain, ytrain)

clf.best_params_, clf.best_score_

Xval = df_val['content'].values
predicted = clf.predict(Xval)
accuracy = accuracy_score(yval, predicted)
accuracy

# Trail with max_feature = 2500
Xtrain = df_train['content']
ytrain = df_train['authId_enc'].values
vectorizer = TfidfVectorizer()
classifierNB = naive_bayes.MultinomialNB()
modelNB = pipeline.Pipeline([("vectorizer", vectorizer),  
                           ("classifier", classifierNB)])
from sklearn.model_selection import GridSearchCV
grid_params = {
  'classifier__alpha': np.linspace(0.5, 1.5, 3),
  'classifier__fit_prior': [True, False],
  'vectorizer__max_df': [0.1, 0.5, 1],
  'vectorizer__binary': [True, False],
  'vectorizer__norm': [None, 'l1', 'l2'], 
  'vectorizer__max_features': [1000, 1500, 2000, 2500,]
}
clf = GridSearchCV(modelNB, grid_params, cv = 2, scoring='accuracy')
clf.fit(Xtrain, ytrain)
print('Best parameter\n', clf.best_params_,'\nBest score\n', clf.best_score_)
Xval = df_val['content'].values
predicted = clf.predict(Xval)
accuracy = accuracy_score(yval, predicted)
accuracy


# Trail with max_feature = 3500
Xtrain = df_train['content']
ytrain = df_train['authId_enc'].values
vectorizer = TfidfVectorizer()
classifierNB = naive_bayes.MultinomialNB()
modelNB = pipeline.Pipeline([("vectorizer", vectorizer),  
                           ("classifier", classifierNB)])
from sklearn.model_selection import GridSearchCV
grid_params = {
  'classifier__alpha': np.linspace(0.5, 1.5, 3),
  'classifier__fit_prior': [True, False],
  'vectorizer__max_df': [0.1, 0.5, 1],
  'vectorizer__binary': [True, False],
  'vectorizer__norm': [None, 'l1', 'l2'], 
  'vectorizer__max_features': [2500,3000, 3500]
}
clf = GridSearchCV(modelNB, grid_params, cv = 2, scoring='accuracy')
clf.fit(Xtrain, ytrain)




print('Best parameter')
clf.best_params_

print('Best score')
clf.best_score_


Xval = df_val['content'].values
predicted = clf.predict(Xval)
accuracy = accuracy_score(yval, predicted)
accuracy

#### chosen model 

df_combined = pd.concat([df_train, df_val], ignore_index=True, axis=0)
df_combined.shape


Xcombined = df_combined['content']
ycombined = df_combined['authId_enc'].values
vectorizer = TfidfVectorizer(max_df=0.5,
                              max_features=3000,
                              norm=None)
classifierNB = naive_bayes.MultinomialNB(fit_prior=False,alpha=0.5)
modelNB_final= pipeline.Pipeline([("vectorizer", vectorizer),  
                           ("classifier", classifierNB)])

modelNB_final.fit(Xcombined, ycombined)

df_test = pd.read_pickle('data/processed/test_clean_df.pkl')

# get lable encoder of author
with open('code/authorIdlabel_3papers.pkl', 'rb') as f:
   authorId_encoder = pickle.load(file = f)


Xtest = df_test['content'].values
predicted = modelNB_final.predict(Xtest)
predictauthorId = authorId_encoder.inverse_transform(predicted)
len(predictauthorId)


predictauthorId

array([  33524946,   33464127, 1390037280, ...,    2814303,    1763912,
         67284811], dtype=int64)

predictauthorId = list(map(str, predictauthorId))
df_test['authorId']= predictauthorId
df_test = df_test[['paperId', 'authorId']].copy()
df_test.head()

df_test.to_json('data/processed/predicted.json', orient="records")

# Testing final model 
Xtrain, Xval = df_train['content'].values, df_val['content'].values
ytrain, yval = df_train['authId_enc'].values, df_val['authId_enc'].values

vectorizer = TfidfVectorizer(max_df=0.5,
                              max_features=3000,
                              norm=None)
classifierNB = naive_bayes.MultinomialNB(fit_prior=False,alpha=0.5)
modelNB_final= pipeline.Pipeline([("vectorizer", vectorizer),  
                           ("classifier", classifierNB)])

modelNB_final.fit(Xtrain, ytrain)


predicted = modelNB_final.predict(Xval)
accuracy = accuracy_score(yval, predicted)
accuracy
