==> code/4.2-prediction.ipynb <==
### Prediction for NLP

import numpy as np
import pandas as pd
from os import getcwd
import glob
from os.path import join
import ntpath
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.feature_extraction.text import CountVectorizer
import os
import pickle

import warnings
import multiprocessing 

def set_path(): 
   abspath = getcwd()
   dname = os.path.dirname(abspath)
   os.chdir(dname)

set_path()
 

def score_NLP(nlpMod, testcase):
    modelcount = len(nlpMod.get_feature_names_out()) #all words in train dictionary of one authorb
    testcount = nlpMod.transform([testcase]).count_nonzero() 
    score = testcount / modelcount
    return score

# prediction for one case:
def prediction(title, abstract):
   # preparing test cases: 
   testcase_NLP = title + " " + abstract

   # access archive:

   pkl_file_list = glob.glob(join(getcwd(),'models_nlp',"*.pkl"))
   result_records = {}
   for fname in pkl_file_list:
      _, authid = ntpath.split(fname)
      authid = int(authid[:-4])
      with open(fname, 'rb') as f:
         model = pickle.load(file = f)
      
      # NLP score
      NLP_model = model
      score_nlp = score_NLP(NLP_model,testcase_NLP)

      result_records[authid] = score_nlp
   return max(result_records, key=result_records.get)
      






df_val= pd.read_pickle('data/processed/val_clean_df.pkl')
df_val.reset_index(inplace=True)
df_val

# predict all 

def predict_all(df):
   pred = []
   for i in range(len(df)):

      
      print(f'case prediction {i}')
      case = df.iloc[i]
      title = case["title"]
      abstract = case["abstract"]
      result = prediction(title, abstract)
      pred.append(result)
   return pred
warnings.filterwarnings('ignore')
val_pred = predict_all(df = df_val)

      

from sklearn.metrics import accuracy_score
accuracy_score(df_val['authorId'], val_pred)
val_pred

df_val.iterrows()

val_pred[3]
