==> code/4.3.prediction.ipynb <==
import numpy as np
import pandas as pd
from os import getcwd
from sklearn import preprocessing
from sklearn.ensemble import GradientBoostingClassifier, RandomForestClassifier
from sklearn.feature_extraction.text import TfidfVectorizer
import os
import pickle
from collections import Counter

def set_path(): 
   abspath = getcwd()
   dname = os.path.dirname(abspath)
   os.chdir(dname)

set_path()
 

df_train = pd.read_pickle('data/processed/train_reduced_df.pkl')
df_val = pd.read_pickle('data/processed/val_reduced_df.pkl')

### Encoding with TFIDF

df_train.head(3)

xtrain, ytrain = df_train['new_content'].values, df_train['authId_enc'].values
xval, yval = df_val['new_content'].values, df_val['authId_enc'].values

tfvectorizer = TfidfVectorizer()
xtrain_tf = tfvectorizer.fit_transform(xtrain)
xval_tf = tfvectorizer.transform(xval)


max(df_train['authId_enc'])

xval_tf.shape,xtrain_tf.shape

yval.shape, ytrain.shape

### Experiment with randomForest

classifier = RandomForestClassifier()

classifier.fit(xtrain_tf,ytrain)

y_pred = classifier.predict(xval_tf)

from sklearn.metrics import accuracy_score
accuracy_score(yval, y_pred)

### MLP

from tensorflow.keras import models
from tensorflow.keras import layers
nb_words = 8000
def build_model():
    ### PUT YOUR CODE HERE ###
    model = models.Sequential()
    model.add(layers.Dense(256, activation='relu', input_shape=(xtrain_tf.shape[1],)))
    model.add(layers.Dense(64, activation='relu'))
    model.add(layers.Dense(5624, activation='sigmoid'))
    model.compile(optimizer='adam',
                  loss='categorical_crossentropy',
                  metrics=['accuracy'])
    return model
    ### END ###

mlp = build_model()
original_hist = mlp.fit(xtrain_tf,
                          ytrain,
                          epochs=20,
                          batch_size=128,
                          validation_data=(xval_tf, yval),
                          verbose=0)
